:shirt: Small fight with pylint here
I could either disable no-member, or I can do this. I can't quite see why
I'd need to do either in this particular case. But, without this, I get
this:

************* Module git2gantt
E: 73,11: Instance of 'PurePath' has no 'exists' member (no-member)

This is odd because of I print the type() of repo out at that point it's
100% not an instance of PurePath.